Refactoring suggestions for ping/readiness on Chapter 1
