+++ src/main.rs
@@ -1,7 +1,7 @@
 mod todo_api_web;
 use todo_api_web::controller::{ping, readiness};
 
-use actix_web::{dev::Service, http::StatusCode, test, web, App, HttpResponse, HttpServer};
+use actix_web::{web, App, HttpResponse, HttpServer};
 
 use num_cpus;
 
@@ -18,20 +18,4 @@ async fn main() -> std::io::Result<()> {
     .unwrap()
     .run()
     .await
-}
-
-#[actix_web::test]
-async fn not_found_route() {
-    let mut app = test::init_service(
-        App::new()
-            .service(readiness)
-            .service(ping)
-            .default_service(web::to(|| HttpResponse::NotFound())),
-    )
-    .await;
-
-    let req = test::TestRequest::with_uri("/crazy-path").to_request();
-
-    let resp = app.call(req).await.unwrap();
-    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
 }
